refactor: DayOfWeek enum 통일
