+++ src/main/java/com/ninedocs/serviceaggregator/client/common/dto/DayOfWeek.java
@@ -1,0 +1,5 @@
+package com.ninedocs.serviceaggregator.client.common.dto;
+
+public enum DayOfWeek {
+  MON, TUE, WED, THU, FRI, SAT, SUN
+}
