--- src/main/java/com/ninedocs/serviceaggregator/controller/mypage/subscription/common/dto/DayOfWeek.java
@@ -1,5 +1,0 @@
-package com.ninedocs.serviceaggregator.controller.mypage.subscription.common.dto;
-
-public enum DayOfWeek {
-    MON, TUE, WED, THU, FRI, SAT, SUN
-  }
