+++ src/main/java/com/ninedocs/serviceaggregator/controller/mypage/subscription/scheduleupsert/dto/UpdateScheduleRequest.java
@@ -1,6 +1,6 @@
 package com.ninedocs.serviceaggregator.controller.mypage.subscription.scheduleupsert.dto;
 
-import com.ninedocs.serviceaggregator.controller.mypage.subscription.common.dto.DayOfWeek;
+import com.ninedocs.serviceaggregator.client.common.dto.DayOfWeek;
 import jakarta.validation.constraints.Size;
 import java.util.List;
 import lombok.Getter;
